% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorTD.R
\name{mirrorTD}
\alias{mirrorTD}
\title{mirrorTD}
\usage{
mirrorTD(varname, td)
}
\arguments{
\item{varname}{Name of active treeio treedata variable tree create}

\item{td}{A treeplyr treedata object}
}
\value{
Creates an Active Binding in the environment the function was run in
}
\description{
Actively Binds a treeio treedata object to a treeplyr 
            treedata object for easier treeplyr interaction with ggtree
}
\examples{
\dontrun{
library(ggtree)
library(treeio)

tree <- fishtree::fishtree_phylogeny(rank="Cyprinidae", 
                                    type="chronogram_mrca")

data <- geiger::sim.char(tree, 0.5, 100) \%>\% matrix() \%>\% as.data.frame()
data$label <- tree$tip.label

td <- make.treedata(tree, data)

mirrorTD(plottingTD, td)

plottingTD \%>\% 
 ggtree(aes(color=V1), 
        layout = "fan") +
 scale_color_continuous(low="lightblue", high="blue") +
 geom_tiplab(color="black", size=.5)
}


}
